[economics] Whitespace move
